refactor(ScalarTypes): remove Generic type, add JSONObject typeBREAKING CHANGE: remove `Generic`, `Json` types, use `JSON` type instead.
