+++ src/type/__tests__/json-test.js
@@ -3,8 +3,7 @@
 // copied from https://github.com/taion/graphql-type-json
 
 import { graphql, GraphQLObjectType, GraphQLSchema } from '../../graphql';
-
-import GraphQLJSON from '../json';
+import { GraphQLJSON } from '..';
 
 const FIXTURE = {
   string: 'string',
@@ -120,7 +119,7 @@ describe('GraphQLJSON', () => {
           value(arg: NaN){
             string: "string"
         }
-       `
+      `
       );
       expect(data).toBeUndefined();
     });
