--- src/type/generic.d.ts
@@ -1,5 +1,0 @@
-import { GraphQLScalarType } from '../graphql';
-
-declare const Generic: GraphQLScalarType;
-
-export default Generic;
