--- src/type/json.d.ts
@@ -1,5 +1,0 @@
-import { GraphQLScalarType } from '../graphql';
-
-declare const JSON: GraphQLScalarType;
-
-export default JSON;
